From Mike Wittman, support for Windows64 build
git-svn-id: http://www.openscenegraph.org/svn/osg/OpenThreads/trunk@5413 16af8721-9629-0410-8352-f15c8da7e697

+++ win32_src/Win32ThreadPrivateData.h
@@ -48,7 +48,7 @@ private:
     Win32ThreadPrivateData() {};
     ~Win32ThreadPrivateData();
 
-    unsigned int stackSize;
+    size_t stackSize;
     bool isRunning;
 
 	int  cancelMode; // 0 - deffered (default) 1-asynch 2-disabled
